Evil 42 add buildings on frontend (#27)
* finish adding events

* cleaning login and register pages

* aligning to new backend

* removing buildingName from floor

* connectiong backend to cycle reservations

* acceot cycle fixed

* deleteRole created

* timer added

* users added to admin panel

* language changed to polish

* small adjustments

* fix build

* buildings added

* fix build

+++ src/app/calendar/_actions/rejectCycleReservation.ts
@@ -1,5 +1,6 @@
 "use server";
 
+import { getBuildingName } from "@/auth/getBuildingName";
 import { getToken } from "@/auth/getToken";
 import { HOST } from "@/server-endpoints/host";
 import { redirect } from "next/navigation";
@@ -8,13 +9,19 @@ export async function rejectCycleReservation(
   recurringId: string,
 ): Promise<void> {
   const token = await getToken();
+  const buildingName = await getBuildingName();
+
+  if (!buildingName) {
+    throw new Error("Building name not found");
+  }
+
   if (!token) {
     console.error("Not authenticated");
     redirect("/login");
   }
 
   const response = await fetch(
-    `${HOST}/buildings/D17/reservations/recurringReservations/${recurringId}/reject`,
+    `${HOST}/buildings/${buildingName}/reservations/recurringReservations/${recurringId}/reject`,
     {
       method: "DELETE",
       headers: {
